LayoutTests/http/wpt/mediarecorder/MediaRecorder-AV-video-only-dataavailable.html

<!doctype html>
<html>
<head>
    <title>MediaRecorder can successfully record the video for a video-only stream</title>
    <link rel="help" href="https://w3c.github.io/mediacapture-record/MediaRecorder.html#mediarecorder">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="../common/canvas-tests.js"></script>
    <link rel="stylesheet" href="../common/canvas-tests.css">
</head>
<body>
<div>
    <video id="player" controls=true>
    </video>
</div>
<div>
    <canvas id="canvas" width="200" height="200">
    </canvas>
    <canvas id="frame" width="200" height="200" style="visibility:visible; display:block">
    </canvas>
</div>
<script>
    var context;
    var drawStartTime;

    function createVideoStream() {
        const canvas = document.getElementById("canvas");
        context = canvas.getContext('2d');
        return canvas.captureStream();
    }

    function doRedImageDraw() {
        if (context) {
            context.fillStyle = "#ff0000";
            context.fillRect(0, 0, 200, 200);
            if (Date.now() - drawStartTime < 2000) {
                window.requestAnimationFrame(doRedImageDraw);
            } else {
                drawStartTime = Date.now();
                doGreenImageDraw();
            }
        }
    }

    function doGreenImageDraw() {
        if (context) {
            context.fillStyle = "#00ff00";
            context.fillRect(0, 0, 200, 200);
            if (Date.now() - drawStartTime < 10000) {
                window.requestAnimationFrame(doGreenImageDraw);
            }
        }
    }

    const video = createVideoStream();
    const recorder = new MediaRecorder(video);
    let mode = 0;
    recorder.ondataavailable = (blobEvent) => {
        assert_true(blobEvent instanceof BlobEvent, 'the type of event should be BlobEvent');
        assert_equals(blobEvent.type, 'dataavailable', 'the event type should be dataavailable');
        assert_true(blobEvent.isTrusted, 'isTrusted should be true when the event is created by C++');
        assert_true(blobEvent.data instanceof Blob, 'the type of data should be Blob');
        assert_true(blobEvent.data.size > 0, 'the blob should contain some buffers');
        player.src = window.URL.createObjectURL(blobEvent.data);
        const resFrame = document.getElementById("frame");
        const resContext = resFrame.getContext('2d');
        player.oncanplay = () => {
            assert_greater_than(player.duration, 0.1, 'the duration should be greater than 100ms');
            player.play();
        };
        player.onplay = () => {
            if (mode === 2)
                return;
            player.pause();
            player.currentTime = 1;
        };
        player.onseeked = () => {
            resContext.drawImage(player, 0, 0);
            if (!mode) {
                _assertPixelApprox(resFrame, 20, 20, 255, 0, 0, 255, "20,2 0", "255, 0, 0, 255", 50);
                _assertPixelApprox(resFrame, 50, 50, 255, 0, 0, 255, "50, 50", "255, 0, 0, 255", 50);
                mode = 1;
                player.currentTime = 3;
            } else if (mode === 1){
                _assertPixelApprox(resFrame, 21, 21, 0, 255, 0, 255, "21, 21", "0, 255, 0, 255", 50);
                _assertPixelApprox(resFrame, 51, 51, 0, 255, 0, 255, "51, 51", "0, 255, 0, 255", 50);
                mode = 2;
                done();
            }
        };
        player.load();
    };
    recorder.start();
    assert_equals(recorder.state, 'recording', 'MediaRecorder has been started successfully');
    drawStartTime = Date.now();
    doRedImageDraw();
    setTimeout(() => {
        recorder.stop();
    }, 4000);

</script>
</body>
</html>
